+++ src/HttpErrorException.php
@@ -62,11 +62,7 @@ class HttpErrorException extends Exception
     /**
      * Create and returns a new instance
      *
-     * @param int            $code     A valid http error code
-     * @param array          $context
-     * @param Throwable|null $previous
-     *
-     * @return static
+     * @param int $code A valid http error code
      */
     public static function create(int $code = 500, array $context = [], Throwable $previous = null): self
     {
@@ -75,22 +71,11 @@ class HttpErrorException extends Exception
         }
 
         $exception = new static(self::$phrases[$code], $code, $previous);
-        $exception->setContext($context);
+        $exception->context = $context;
 
         return $exception;
     }
 
-    /**
-     * Add data context used in the error handler
-     */
-    public function setContext(array $context)
-    {
-        $this->context = $context;
-    }
-
-    /**
-     * Return the data context
-     */
     public function getContext(): array
     {
         return $this->context;